+++ pkg/license/logger.go
@@ -18,36 +18,10 @@
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
 //
 
-package assertion
+package license
 
-import (
-	"fmt"
+import "github.com/arangodb/kube-arangodb/pkg/logging"
+
+var (
+	logger = logging.Global().RegisterAndGetLogger("license", logging.Info)
 )
-
-type Key string
-
-const (
-	KeyUnknown               Key = ""
-	DeprecatedActionKey      Key = "DeprecatedAction"
-	CommunityLicenseCheckKey Key = "CommunityLicenseCheck"
-)
-
-func (k Key) Assert(condition bool, msg string, args ...interface{}) {
-	assert(2, condition, k, msg, args...)
-}
-
-func Assert(condition bool, key Key, msg string, args ...interface{}) {
-	assert(2, condition, key, msg, args...)
-}
-
-func assert(skip int, condition bool, key Key, msg string, args ...interface{}) {
-	if !condition {
-		return
-	}
-
-	metricsObject.incKeyMetric(key)
-
-	frames := frames(skip)
-
-	_assert(frames, fmt.Sprintf(msg, args...))
-}